Refactors data api to address circular dependencies.
