+++ assets/js/googlesitekit/datastore/site/index.test.js
@@ -29,10 +29,8 @@ import apiFetch from '@wordpress/api-fetch';
  * Internal dependencies
  */
 import { createTestRegistry, unsubscribeFromAll } from 'tests/js/utils';
-import {
-	INITIAL_STATE,
-	STORE_NAME,
-} from './index';
+import { INITIAL_STATE } from './index';
+import { STORE_NAME } from './constants';
 
 describe( 'core/site store', () => {
 	let apiFetchSpy;
