+++ assets/js/googlesitekit/datastore/site/info.test.js
@@ -25,7 +25,8 @@ import {
 	subscribeUntil,
 	unsubscribeFromAll,
 } from 'tests/js/utils';
-import { INITIAL_STATE, STORE_NAME } from './index';
+import { INITIAL_STATE } from './index';
+import { STORE_NAME } from './constants';
 
 describe( 'core/site site info', () => {
 	const baseInfoVar = '_googlesitekitBaseData';
